feat(utils)!: We updated choose() to remove types.
BREAKING CHANGE: We updated the choose() function, so it doesn't need generic types anymore, it will infer the return type now from the first .case() or .type() call. As a consequence, the out property is now ONLY a function, not an object anymore - we needed to do this to prevent edge cases with the use of any's.