feat(easy): Renewed the `View` class.
BREAKING CHANGE: The `View` class does not use the `{in, col}` notation anymore.

The `{in, col}` notation in views was originally meant to make the view go 2-ways. However, in real use, we never do that. So, I simplified it, which also makes it easier to understand, and to extend. Replace the `{in, col}` with functions. You can now also use constructors and other views directly.