feat: Add 'dir' function to Log and its corresponding test
This commit includes an additional 'dir' function in the 'Log.ts' module, which is designed to display the properties of any JavaScript object which is passed as a parameter. A test has also been added to the 'Log.test.ts' module to verify the correct working of the newly introduced 'dir' function.