feat: Update luxon library and improve code consistency
This commit primarily includes an update of the version of the 'luxon' library from 3.3.0 to 3.4.4, and corresponding '@types/luxon' from 3.2.0 to 3.4.2. Minor refactoring was also completed in several files to improve readability and please note the adjustment to the syntax in 'DateTime.ts'. The security handler, in 'SecurityHandler.ts', was adjusted for easier readability and slight changes were made to 'Identity.ts' and 'Scope.ts'.